@@ -31,24 +31,25 @@
     "barcode-detector": "^0.5.0",
     "callforth": "^0.3.1",
     "core-js": "^3.6.5",
-    "vue": "^2.6.11",
+    "vue": "^3.0.0",
     "webrtc-adapter": "7.7.0"
   },
   "devDependencies": {
-    "@vue/cli-plugin-babel": "~4.4.0",
-    "@vue/cli-plugin-eslint": "~4.4.0",
-    "@vue/cli-service": "~4.4.0",
+    "@vue/cli-plugin-babel": "^4.5.8",
+    "@vue/cli-plugin-eslint": "^4.5.8",
+    "@vue/cli-service": "^4.5.8",
+    "@vue/compiler-sfc": "^3.0.0",
     "@vue/eslint-config-prettier": "^6.0.0",
+    "@vue/eslint-config-standard": "^5.1.2",
     "babel-eslint": "^10.1.0",
     "eslint": "^6.7.2",
     "eslint-plugin-prettier": "^3.1.3",
-    "eslint-plugin-vue": "^6.2.2",
+    "eslint-plugin-vue": "^7.0.0-0",
     "husky": "^4.2.5",
     "lint-staged": "^9.5.0",
     "prettier": "^1.19.1",
     "semantic-release": "^17.1.1",
-    "vue-template-compiler": "^2.6.11",
-    "vuepress": "^1.5.2"
+    "vuepress": "^1.7.1"
   },
   "gitHooks": {
     "pre-commit": "lint-staged"
